Refactor(payment): 결제 도메인 공통 타입을 common 모듈로 이동 및 의존성 정리
- PaymentStatus, PaymentType, RefundStatus를 common 모듈의 shared 패키지로 이동
- 도메인 간 결합도를 낮추고 결제 상태 정보를 공통으로 관리할 수 있도록 개선
- 관련 엔티티 및 DTO의 import 참조 업데이트
