--- semicolon/src/main/java/dukku/semicolon/boundedContext/payment/entity/enums/PaymentType.java
@@ -1,9 +1,0 @@
-package dukku.semicolon.boundedContext.payment.entity.enums;
-
-/**
- * 결제 유형
- */
-public enum PaymentType {
-    DEPOSIT, // 예치금 결제
-    MIXED // 복합 결제 (예치금 + PG)
-}
